fixed Scrum Report Page and refine the History Description related is…

+++ java/ntut/csie/ezScrum/web/control/ScheduleReport.java
@@ -56,7 +56,7 @@ public class ScheduleReport {
 	private String mSprintGoal = "";
 	private int mSize = 0;
 
-	public ScheduleReport(ProjectObject project, IUserSession session, int iteration) {
+	public ScheduleReport(ProjectObject project, IUserSession session, long iteration) {
 		mProject = project;
 		mIteration = iteration;
 	}
